# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/invitations": {
      "post": {
        "description": "Admin: companyId y role (business|common) obligatorios. Business: invita solo common a su empresa.",
        "operationId": "InvitationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvitationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Sin permisos"
          },
          "409": {
            "description": "Email o invitación pendiente duplicada"
          },
          "422": {
            "description": "Faltan companyId/role o el actor no tiene empresa"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Invitar usuario",
        "tags": [
          "Invitaciones"
        ]
      },
      "get": {
        "operationId": "InvitationsController_findAll",
        "parameters": [
          {
            "name": "companyId",
            "required": false,
            "in": "query",
            "description": "Filtro opcional (solo admin)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvitationResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Sin permisos"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Listar invitaciones pendientes",
        "tags": [
          "Invitaciones"
        ]
      }
    },
    "/api/invitations/{id}": {
      "delete": {
        "operationId": "InvitationsController_revoke",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Revocada"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Sin permisos"
          },
          "404": {
            "description": "Invitación no encontrada"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Revocar invitación pendiente",
        "tags": [
          "Invitaciones"
        ]
      }
    },
...
    "/api/quotations": {
      "get": {
        "description": "Devuelve todas las cotizaciones de la empresa del usuario, ordenadas por fecha de creación descendente.",
        "operationId": "QuotationsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuotationResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Listar cotizaciones de la empresa",
        "tags": [
          "Cotizaciones"
        ]
      },
      "post": {
        "description": "Crea una cotización para la empresa del usuario autenticado. Los subtotales y el total se recalculan en el servidor.",
        "operationId": "QuotationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuotationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotationResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "422": {
            "description": "El usuario no tiene empresa configurada"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Crear cotización",
        "tags": [
          "Cotizaciones"
        ]
      }
    },
    "/api/quotations/{id}/pdf": {
      "get": {
        "description": "Genera un PDF con el mismo estilo que la vista de cotización. Requiere datos de empresa configurados (PUT /api/company).",
        "operationId": "QuotationsController_downloadPdf",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Archivo PDF (descarga)"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "404": {
            "description": "Cotización no encontrada"
          },
          "422": {
            "description": "Datos de empresa no configurados; configure con PUT /api/company"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Descargar PDF de la cotización",
        "tags": [
          "Cotizaciones"
        ]
      }
    },
    "/api/quotations/{id}": {
      "get": {
        "description": "Devuelve la cotización si pertenece a la empresa del usuario autenticado.",
        "operationId": "QuotationsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotationResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "404": {
            "description": "Cotización no encontrada"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Obtener cotización por id",
        "tags": [
          "Cotizaciones"
        ]
      },
      "put": {
        "description": "Reemplaza la cotización completa (incluyendo items) si pertenece a la empresa del usuario autenticado.",
        "operationId": "QuotationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuotationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotationResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "404": {
            "description": "Cotización no encontrada"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Actualizar cotización por id",
        "tags": [
          "Cotizaciones"
        ]
      }
    },
    "/api/quotations/{id}/send": {
      "post": {
        "description": "Genera el PDF, lo envía al clientEmail con plantilla Handlebars vía Resend (SMTP) y marca la cotización como `sent` si estaba en `draft`. Permite reenvío si ya está en `sent`.",
        "operationId": "QuotationsController_sendByEmail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotationResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "404": {
            "description": "Cotización no encontrada"
          },
          "409": {
            "description": "Estado no permite envío (approved, rejected o expired)"
          },
          "422": {
            "description": "clientEmail ausente o empresa no configurada"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Enviar cotización por correo al cliente",
        "tags": [
          "Cotizaciones"
        ]
      }
    },
    "/api/quotations/{id}/status": {
      "patch": {
        "description": "Solo cotizaciones en estado `sent` y vigentes pueden pasar a `approved` o `rejected`.",
        "operationId": "QuotationsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuotationStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotationResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "404": {
            "description": "Cotización no encontrada"
          },
          "409": {
            "description": "No está en sent o ya está expirada"
          },
          "422": {
            "description": "status distinto de approved o rejected"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Actualizar estado de la cotización (aprobada o rechazada)",
        "tags": [
          "Cotizaciones"
        ]
      }
    },
...
  },
  "info": {
    "title": "Cotizador API",
    "description": "Contrato OpenAPI 3 para el front (`cotizador-app`).\n\nPrefijo global: `/api`. Auth: header `Authorization: Bearer <jwt>` (scheme `access-token`).\nJSON vivo: `GET /api/openapi.json`. UI: `/api/docs`.\nGuía de pantallas, roles y labels: `docs/contrato-front.md`.\n\nErrores: `{ statusCode, message, traceId? }`. `message` puede ser `string` o `string[]` (validación).\nMontos en CLP (número). Fechas de negocio `YYYY-MM-DD`; timestamps ISO-8601.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Token obtenido en POST /api/auth/login"
      }
    },
    "schemas": {
...
      "CreateInvitationDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "nuevo@empresa.cl"
          },
          "name": {
            "type": "string",
            "example": "Ana Pérez"
          },
          "companyId": {
            "type": "string",
            "format": "uuid",
            "description": "Obligatorio para admin. Business lo ignora (usa su empresa)."
          },
          "role": {
            "type": "string",
            "enum": [
              "business",
              "common"
            ],
            "description": "Obligatorio para admin. Business siempre invita common (se ignora)."
          }
        },
        "required": [
          "email",
          "name"
        ]
      },
      "InvitationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "business",
              "common"
            ]
          },
          "companyId": {
            "type": "string",
            "format": "uuid"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "acceptedAt": {
            "type": "object",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role",
          "companyId",
          "expiresAt",
          "acceptedAt",
          "createdAt"
        ]
      },
...
      "QuotationItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1777653994936-0"
          },
          "description": {
            "type": "string",
            "example": "Mano de obra"
          },
          "quantity": {
            "type": "number",
            "example": 1
          },
          "unitPrice": {
            "type": "number",
            "example": 2000000
          },
          "subtotal": {
            "type": "number",
            "example": 2000000
          }
        },
        "required": [
          "id",
          "description",
          "quantity",
          "unitPrice",
          "subtotal"
        ]
      },
      "QuotationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "clientName": {
            "type": "string",
            "example": "Constructora Carlos S.A"
          },
          "clientRut": {
            "type": "object",
            "example": "76.349.234-1",
            "nullable": true
          },
          "clientEmail": {
            "type": "object",
            "example": "factura@constructoracarlos.cl",
            "nullable": true
          },
          "projectTitle": {
            "type": "object",
            "example": "Renovación Casa",
            "nullable": true
          },
          "projectDeadline": {
            "type": "object",
            "example": "2026-12-05",
            "nullable": true
          },
          "projectNotes": {
            "type": "object",
            "example": "",
            "nullable": true
          },
          "validUntil": {
            "type": "object",
            "example": "2026-06-30",
            "nullable": true,
            "description": "Validez de la oferta (solo fecha)."
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "sent",
              "approved",
              "rejected",
              "expired"
            ],
            "example": "draft"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuotationItemResponseDto"
            }
          },
          "total": {
            "type": "number",
            "example": 2000000
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "clientName",
          "clientRut",
          "clientEmail",
          "projectTitle",
          "projectDeadline",
          "projectNotes",
          "validUntil",
          "status",
          "items",
          "total",
          "createdAt",
          "updatedAt"
        ]
      },
      "QuotationItemInputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identificador del item generado por el cliente. Si se omite, el servidor genera uno.",
            "example": "1777653994936-0"
          },
          "description": {
            "type": "string",
            "example": "Mano de obra"
          },
          "quantity": {
            "type": "number",
            "example": 1,
            "minimum": 1
          },
          "subtotal": {
            "type": "number",
            "example": 1,
            "minimum": 1
          },
          "unitPrice": {
            "type": "number",
            "example": 2000000,
            "minimum": 0
          }
        },
        "required": [
          "description",
          "quantity",
          "subtotal",
          "unitPrice"
        ]
      },
      "CreateQuotationDto": {
        "type": "object",
        "properties": {
          "clientName": {
            "type": "string",
            "example": "Constructora Carlos S.A"
          },
          "clientRut": {
            "type": "string",
            "example": "76.349.234-1"
          },
          "clientEmail": {
            "type": "string",
            "example": "factura@constructoracarlos.cl"
          },
          "projectTitle": {
            "type": "string",
            "example": "Renovación Casa"
          },
          "projectDeadline": {
            "type": "string",
            "example": "2026-12-05",
            "description": "Fecha límite del proyecto (string libre, puede venir vacío)."
          },
          "projectNotes": {
            "type": "string",
            "example": ""
          },
          "validUntil": {
            "type": "string",
            "example": "2026-06-30",
            "description": "Validez de la oferta (solo fecha ISO). Si la fecha ya pasó y el estado almacenado es borrador o enviado, la API expondrá el estado `expired`."
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuotationItemInputDto"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "sent",
              "approved",
              "rejected",
              "expired"
            ],
            "example": "draft",
            "default": "draft"
          }
        },
        "required": [
          "clientName",
          "items"
        ]
      },
      "UpdateQuotationStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "approved",
              "rejected"
            ],
            "example": "approved"
          }
        },
        "required": [
          "status"
        ]
      },
      "UpdateQuotationDto": {
        "type": "object",
        "properties": {
          "clientName": {
            "type": "string",
            "example": "Constructora Carlos S.A"
          },
          "clientRut": {
            "type": "string",
            "example": "76.349.234-1"
          },
          "clientEmail": {
            "type": "string",
            "example": "factura@constructoracarlos.cl"
          },
          "projectTitle": {
            "type": "string",
            "example": "Renovación Casa"
          },
          "projectDeadline": {
            "type": "string",
            "example": "2026-12-05",
            "description": "Fecha límite del proyecto (string libre, puede venir vacío)."
          },
          "projectNotes": {
            "type": "string",
            "example": ""
          },
          "validUntil": {
            "type": "string",
            "example": "2026-06-30",
            "description": "Validez de la oferta (solo fecha ISO). Si la fecha ya pasó y el estado almacenado es borrador o enviado, la API expondrá el estado `expired`."
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuotationItemInputDto"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "sent",
              "approved",
              "rejected",
              "expired"
            ],
            "example": "draft",
            "default": "draft"
          }
        },
        "required": [
          "clientName",
          "items"
        ]
      },
...
    }
  }
}