# src/quotations/quotations.controller.ts
import {
  Body,
  Controller,
  Get,
  HttpCode,
  HttpStatus,
  Param,
  ParseUUIDPipe,
  Patch,
  Post,
  Put,
  Req,
  Res,
  StreamableFile,
  UseGuards,
} from '@nestjs/common';
import {
  ApiBearerAuth,
  ApiBody,
  ApiConflictResponse,
  ApiCreatedResponse,
  ApiNotFoundResponse,
  ApiOkResponse,
  ApiOperation,
  ApiProduces,
  ApiTags,
  ApiUnauthorizedResponse,
  ApiUnprocessableEntityResponse,
} from '@nestjs/swagger';
import type { Request, Response } from 'express';
import { JwtAuthGuard } from '../auth/jwt-auth.guard';
import type { JwtPayload } from '../auth/jwt-payload.type';
import { Roles } from '../auth/roles.decorator';
import { RolesGuard } from '../auth/roles.guard';
import { CompanyService } from '../company/company.service';
import { CompanyTermsService } from '../company/company-terms.service';
import { CreateQuotationDto } from './dto/create-quotation.dto';
import { QuotationResponseDto } from './dto/quotation-response.dto';
import { UpdateQuotationDto } from './dto/update-quotation.dto';
import { UpdateQuotationStatusDto } from './dto/update-quotation-status.dto';
import { buildQuoteNumber } from './pdf/quotation-pdf.formatters';
import { QuotationPdfService } from './pdf/quotation-pdf.service';
import { QuotationsService } from './quotations.service';

type RequestWithUser = Request & { user: JwtPayload };

@ApiTags('Cotizaciones')
@ApiBearerAuth('access-token')
@UseGuards(JwtAuthGuard, RolesGuard)
@Roles('business', 'common')
@Controller('quotations')
export class QuotationsController {
  constructor(
    private readonly quotationsService: QuotationsService,
    private readonly companyService: CompanyService,
    private readonly companyTermsService: CompanyTermsService,
    private readonly quotationPdfService: QuotationPdfService,
  ) {}

  @Get()
  @ApiOperation({
    summary: 'Listar cotizaciones de la empresa',
    description:
      'Devuelve todas las cotizaciones de la empresa del usuario, ordenadas por fecha de creación descendente.',
  })
  @ApiOkResponse({ type: [QuotationResponseDto] })
  @ApiUnauthorizedResponse({ description: 'Token ausente o inválido' })
  findAll(@Req() req: RequestWithUser): Promise<QuotationResponseDto[]> {
    return this.quotationsService.findAllByUser(req.user.sub);
  }

  @Get(':id/pdf')
  @ApiOperation({
    summary: 'Descargar PDF de la cotización',
    description:
      'Genera un PDF con el mismo estilo que la vista de cotización. Requiere datos de empresa configurados (PUT /api/company).',
  })
  @ApiProduces('application/pdf')
  @ApiOkResponse({ description: 'Archivo PDF (descarga)' })
  @ApiUnauthorizedResponse({ description: 'Token ausente o inválido' })
  @ApiNotFoundResponse({ description: 'Cotización no encontrada' })
  @ApiUnprocessableEntityResponse({
    description:
      'Datos de empresa no configurados; configure con PUT /api/company',
  })
  async downloadPdf(
    @Req() req: RequestWithUser,
    @Param('id', new ParseUUIDPipe()) id: string,
    @Res({ passthrough: true }) res: Response,
  ): Promise<StreamableFile> {
    const company = await this.companyService.findByUserOrFail(req.user.sub);
    const quotation = await this.quotationsService.findOneByUser(
      req.user.sub,
      id,
    );
    const terms = await this.companyTermsService.resolveTermsForUser(
      req.user.sub,
    );
    const buffer = await this.quotationPdfService.generate(
      quotation,
      this.companyService.toResponse(company),
      terms,
      req.user.sub,
    );
    const quoteNumber = buildQuoteNumber(quotation.id);
    res.set({
      'Content-Type': 'application/pdf',
      'Content-Disposition': `attachment; filename="cotizacion-${quoteNumber}.pdf"`,
    });
    return new StreamableFile(buffer);
  }

  @Get(':id')
  @ApiOperation({
    summary: 'Obtener cotización por id',
    description:
      'Devuelve la cotización si pertenece a la empresa del usuario autenticado.',
  })
  @ApiOkResponse({ type: QuotationResponseDto })
  @ApiUnauthorizedResponse({ description: 'Token ausente o inválido' })
  @ApiNotFoundResponse({ description: 'Cotización no encontrada' })
  findOne(
    @Req() req: RequestWithUser,
    @Param('id', new ParseUUIDPipe()) id: string,
  ): Promise<QuotationResponseDto> {
    return this.quotationsService.findOneByUser(req.user.sub, id);
  }

  @Post()
  @HttpCode(HttpStatus.CREATED)
  @ApiOperation({
    summary: 'Crear cotización',
    description:
      'Crea una cotización para la empresa del usuario autenticado. Los subtotales y el total se recalculan en el servidor.',
  })
  @ApiBody({ type: CreateQuotationDto })
  @ApiCreatedResponse({ type: QuotationResponseDto })
  @ApiUnauthorizedResponse({ description: 'Token ausente o inválido' })
  @ApiUnprocessableEntityResponse({
    description: 'El usuario no tiene empresa configurada',
  })
  create(
    @Req() req: RequestWithUser,
    @Body() dto: CreateQuotationDto,
  ): Promise<QuotationResponseDto> {
    return this.quotationsService.create(req.user.sub, dto);
  }

  @Post(':id/send')
  @HttpCode(HttpStatus.OK)
  @ApiOperation({
    summary: 'Enviar cotización por correo al cliente',
    description:
      'Genera el PDF, lo envía al clientEmail con plantilla Handlebars vía Resend (SMTP) y marca la cotización como `sent` si estaba en `draft`. Permite reenvío si ya está en `sent`.',
  })
  @ApiOkResponse({ type: QuotationResponseDto })
  @ApiUnauthorizedResponse({ description: 'Token ausente o inválido' })
  @ApiNotFoundResponse({ description: 'Cotización no encontrada' })
  @ApiConflictResponse({
    description: 'Estado no permite envío (approved, rejected o expired)',
  })
  @ApiUnprocessableEntityResponse({
    description: 'clientEmail ausente o empresa no configurada',
  })
  sendByEmail(
    @Req() req: RequestWithUser,
    @Param('id', new ParseUUIDPipe()) id: string,
  ): Promise<QuotationResponseDto> {
    return this.quotationsService.sendByEmail(req.user.sub, id);
  }

  @Patch(':id/status')
  @ApiOperation({
    summary: 'Actualizar estado de la cotización (aprobada o rechazada)',
    description:
      'Solo cotizaciones en estado `sent` y vigentes pueden pasar a `approved` o `rejected`.',
  })
  @ApiBody({ type: UpdateQuotationStatusDto })
  @ApiOkResponse({ type: QuotationResponseDto })
  @ApiUnauthorizedResponse({ description: 'Token ausente o inválido' })
  @ApiNotFoundResponse({ description: 'Cotización no encontrada' })
  @ApiConflictResponse({
    description: 'No está en sent o ya está expirada',
  })
  @ApiUnprocessableEntityResponse({
    description: 'status distinto de approved o rejected',
  })
  updateStatus(
    @Req() req: RequestWithUser,
    @Param('id', new ParseUUIDPipe()) id: string,
    @Body() dto: UpdateQuotationStatusDto,
  ): Promise<QuotationResponseDto> {
    return this.quotationsService.updateStatus(req.user.sub, id, dto);
  }

  @Put(':id')
  @ApiOperation({
    summary: 'Actualizar cotización por id',
    description:
      'Reemplaza la cotización completa (incluyendo items) si pertenece a la empresa del usuario autenticado.',
  })
  @ApiBody({ type: UpdateQuotationDto })
  @ApiOkResponse({ type: QuotationResponseDto })
  @ApiUnauthorizedResponse({ description: 'Token ausente o inválido' })
  @ApiNotFoundResponse({ description: 'Cotización no encontrada' })
  update(
    @Req() req: RequestWithUser,
    @Param('id', new ParseUUIDPipe()) id: string,
    @Body() dto: UpdateQuotationDto,
  ): Promise<QuotationResponseDto> {
    return this.quotationsService.update(req.user.sub, id, dto);
  }
}